+++ src/components/regions/MainSelectors.svelte
@@ -1,13 +1,15 @@
 <script>
-import MenuButton from 'udgl/menu/MenuButton.svelte';
+import { fly } from 'svelte/transition';
 import MenuList from 'udgl/menu/MenuList.svelte';
 import MenuListItem from 'udgl/menu/MenuListItem.svelte';
 import DownCarat from 'udgl/icons/DownCarat.svelte';
-import CONFIG from '../../config/firefox-desktop';
+import DimensionMenu from '../controls/DimensionMenu.svelte';
+import productConfig from '../../config/products';
 
 import {
   store,
   getFieldValueLabel,
+  probe,
 } from '../../state/store';
 
 const OFFSET = 10;
@@ -29,6 +31,7 @@ const COMPACT = true;
   grid-auto-flow: column;
   align-items: center;
   grid-column-gap: var(--space-1h);
+  font-weight: 500;
 }
 
 .pull-right-edge {
@@ -37,53 +40,43 @@ const COMPACT = true;
   align-items: center;
 }
 
+.main-filter__label__dimension {
+  text-transform: uppercase;
+  padding-right: var(--space-1q);
+  color: var(--digital-blue-600);
+  font-weight: 300;
+}
+
+.main-filter__label__dimension:after {
+  content: " / ";
+  padding-left: var(--space-1q);
+  color: var(--cool-gray-400);
+}
+
 </style>
 
-<div class='main-filters'>
-  <MenuButton tooltip={'Select a Channel'} compact={COMPACT} offset={OFFSET} location='bottom' alignment='right'>
-    <div class=main-filter__label slot="label">{getFieldValueLabel('channel', $store.productDimensions.channel)} <div class=pull-right-edge><DownCarat size=14 /></div></div>
-    <div slot="menu">
-      <MenuList on:selection={(event) => { store.setDimension('channel', event.detail.key); }}>
-          {#each CONFIG.dimensions.channel.values as {key, label}, i (key)}
-            <MenuListItem  key={key} value={key}><span class='story-label
-              first'></span>{label}</MenuListItem>
-            {/each}
-        </MenuList>
-    </div>
-  </MenuButton>
-  <MenuButton tooltip={'Select an OS'} compact={COMPACT} offset={OFFSET} location='bottom' alignment='right'>
-    <div class=main-filter__label slot="label">{getFieldValueLabel('os', $store.productDimensions.os)}<div class=pull-right-edge><DownCarat size=14 /></div></div>
-    <div slot="menu">
-      <MenuList on:selection={(event) => { store.setDimension('os', event.detail.key); }}>
-        {#each CONFIG.dimensions.os.values as {key, label}, i (key)}
-          <MenuListItem key={key} value={key}><span class='story-label
-            first'></span>{label}</MenuListItem>
-          {/each}
-      </MenuList>
-    </div>
-  </MenuButton>
-  <MenuButton tooltip={'Select an Aggregation Level'} compact={COMPACT} offset={OFFSET} location='bottom' alignment='right'>
-    <div class=main-filter__label slot="label">{getFieldValueLabel('aggregationLevel', $store.productDimensions.aggregationLevel)}<div class=pull-right-edge><DownCarat size=14 /></div></div>
-    <div slot="menu">
-        <MenuList on:selection={(event) => { store.setDimension('aggregationLevel', event.detail.key); }}>
-          {#each CONFIG.dimensions.aggregationLevel.values as {key, label}, i (key)}
-          <MenuListItem key={key} value={key}><span class='story-label
-            first'></span>{label}</MenuListItem>
-          {/each}
-        </MenuList>
-    </div>
-  </MenuButton>
-  {#if $store.route.view} <!-- Hide process selector on home page. -->
-    <MenuButton tooltip={'Select a Process'} compact={COMPACT} offset={OFFSET} location='bottom' alignment='left'>
-      <div class=main-filter__label slot="label">{getFieldValueLabel('process', $store.productDimensions.process) || 'select a process'}<div class=pull-right-edge><DownCarat size=14 /></div></div>
-      <div slot="menu">
-          <MenuList on:selection={(event) => { store.setDimension('process', event.detail.key); }}>
-            {#each CONFIG.dimensions.process.values as {key, label}, i (key)}
-            <MenuListItem  key={key} value={key}><span class='story-label
-              first'></span>{label}</MenuListItem>
-            {/each}
-          </MenuList>
-      </div>
-    </MenuButton>
-  {/if}
-</div>
+{#if $store.route.section === 'probe' && $probe}
+  <div transition:fly={{ x: 5, duration: 200 }} class='main-filters'>
+    {#each Object.values(productConfig[$store.product].dimensions) as dimension, i (dimension.key)}
+      {#if dimension.values.some(
+            (di) => dimension.isValidKey === undefined
+                    || dimension.isValidKey(di.key, $probe, store),
+      )}
+      <DimensionMenu tooltip='Select a {dimension.title}' compact={COMPACT} offset={OFFSET} location='bottom' alignment='right'>
+        <div class=main-filter__label slot="label">
+          <span class='main-filter__label__dimension'>{dimension.title}</span>
+          {getFieldValueLabel(dimension.key, $store.productDimensions[dimension.key])} <div class=pull-right-edge><DownCarat size=14 /></div></div>
+        <div slot="menu">
+          <MenuList on:selection={(event) => { store.setDimension(dimension.key, event.detail.key); }}>
+              {#each dimension.values.filter((di) => dimension.isValidKey === undefined
+                || dimension.isValidKey(di.key, $probe, store)) as {key, label}, i (key)}
+                <MenuListItem  key={key} value={key}><span class='story-label
+                  first'></span>{label}</MenuListItem>
+                {/each}
+            </MenuList>
+        </div>
+      </DimensionMenu>
+      {/if}
+    {/each}
+  </div>
+{/if}